restructuration du projet en classes
