+++ .gitignore
@@ -1,7 +1,7 @@
+__pycache__/
 _archives/
 .vscode/
 env/
 img/_archives/
+src/__pycache__/
 aide.txt
-config.py
-hero.py
