+++ config/configs.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+CELL_SIZE = 50
+SCREEN_SIZE = (750, 750)
+BLACK = (0, 0, 0)
+LANE_POINT = 'C'
+WALL_POINT = '#'
+START_POINT = 'S'
+EXIT_POINT = 'E'
