+++ src/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .hero import Hero
+from .point import Point
+from .game_map import GameMap
+from .game import Game
